+++ src/directives/index.ts
@@ -1,0 +1,5 @@
+import { DebugDirective } from './debug'
+
+export const schemaDirectives = {
+  debug: DebugDirective,
+}